Initial commit for band pass de/modulator.
Implemented band pass modulator & demodulator, this interleaves the samples in following fashion and places into appropriate buffer, output array for the encoder and fft input on for the decoder.

Added unit test and modified integration test to include new features. All tests pass.

TODO: Currently the modulator and demodulator works for even size Fourier transform greater than 4. New methods need to be implemented to handle all sizes of the fourier transforms.

Notes: The performance of the modulator and demodulator  needs to be assessed.

Modification of other objects:

The FFT plan type was changed to measure the performance and choose appropriate method, initial investigation proved that the transform execution time in both directions decreased.

A futrhter investigation is necessary to verify the fastest fft plan.

File structure:

Added new folder for band pass de/modulator.

Changed the namer of ofdmcoder to ofdmcodec.