Initial commit on the cyclic prefix feature
Added correlator object which  works in the time domain and assosiated unit test.

Temporarily using gnuplot for plotting waveforms.

Notes:
The cyclic prefix addition on the tx side has not yet been implemented, for now a crude method of copying the memory to the array is used for test purposes.

The correlation execution time for a whole block of data (in the test 20 upsampled symbols with prefix) is big in the range of 10 miliseconds.

Additional logic for thersholding is expected to increase this time, therefore an evaluation of other methods should be carried out.

There is a lot of outstanding work to be done for the release 0.3