Correlator Update
Cyclic-Prefix:
Cyclic-prefix is a standalone function which copies the end portion of the symbol to its front, this function takes the pointer to the start of the prefix as an input parameter.

Coarse Search remains in time domain, the feasability of the frequency domain methods is going to be assesed during or after audio driver release.

Flipping the prefix was uneccesary, this patch rectifies this bug.

Now the correlator output on the Rx datablock is compared against the object's threshold. A function needs to be implemented to either set this level by listening to the noise when nothing is being transmitted or switch the logic to peak detection.

Added cyclic prefix unit test and updated other unit and integration tests to reflect all changes.

Band-Pass Modulator:
Now band pass modulator & demodulator takes the pointer to the start of the symbol.

FFT Object:
Improved the encoding process by adding function which uses initalized fftw plan to save output to specified destination. This removes the need to copy the whole symbol, which was  was done when symbol was upsampled in band pass modulator.