Added QAM Modulator
Modified all relevant features to reflect this addition.

Added unit test & modified other tests to refelect the changes.

Note:
This QAM modulator is a fast hack that is not so efficient, ideally the modulation/demodulation should be handled in individual loops to decreas complexity & execution time, this exteends to fft's sum of imag function.