Created Transcievier object
Trx object wraps all required functions for transmitting and receving audio in real-time and also encapsulates ofdm codec which is going to be used in callbacks in the future.

Note: this is a simple example which plays back an audio from raw file, the recording functionality is not yet implemented.